requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.7
django-admin-rangefilter==0.9.0
django-appconf==1.0.5
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-compressor==4.0
django-crispy-forms==1.12.0
django-htmlmin==0.11.0
django-storages==1.12.3
django-storages-redux==1.3.3
django-timezone-field==5.0
docutils==0.17.1
gunicorn==20.1.0
html5lib==1.1
idna==3.4
jmespath==1.0.0
kombu==5.2.4
packaging==21.3
Pillow==8.3.2
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-memcached==1.59
python3-openid==3.2.0
pytz==2022.7.1
rcssmin==1.1.0
redis==4.3.4
requests==2.26.0
rjsmin==1.2.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
toml==0.10.2
twilio==7.16.4
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
wrapt==1.15.0
